# lib/swagger-model.yml
---
swagger: "2.0"
info:
  description: "This is the swagger model for the throw up the X Fantasy Football\
    \ Viewer\n"
  version: "1.0.0"
  title: "ThrowUpTheX"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/William-Bulovas/ThrowUpTheX/1.0.0"
tags:
- name: "draftResults"
  description: "Gets the draft results for past seasons"
schemes:
- "https"
- "http"
paths:
  /draft/findByManager:
    get:
      tags:
      - "draftResults"
      summary: "Finds Pets by status"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findPetsByStatus"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values that need to be considered for filter"
        required: true
        type: "array"
        items:
          type: "string"
          enum:
          - "available"
          - "pending"
          - "sold"
          default: "available"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DraftResult"
  /draft/findByYear:
    get:
      tags:
      - "draftResults"
      summary: "Finds Pets by tags"
      description: "Muliple tags can be provided with comma separated strings. Use\\\
        \ \\ tag1, tag2, tag3 for testing."
      operationId: "findPetsByTags"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "tags"
        in: "query"
        description: "Tags to filter by"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DraftResult"
        "400":
          description: "Invalid tag value"
  /draft/{managerId}/{year}:
    get:
      tags:
      - "draftResults"
      summary: "Find pet by ID"
      description: "Returns the draft results for a year and manager"
      operationId: "getPetById"
      produces:
      - "application/json"
      parameters:
      - name: "managerId"
        in: "path"
        description: "Manager of the draft"
        required: true
        type: "string"
      - name: "year"
        in: "path"
        description: "Fantasy Football year"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DraftResult"
        "404":
          description: "Manager/Year not found"
definitions:
  DraftResult:
    type: "object"
    properties:
      managerId:
        type: "string"
      managerName:
        type: "string"
      teamId:
        type: "string"
      teamName:
        type: "string"
      playerId:
        type: "string"
      playerName:
        type: "string"
      playerPosition:
        type: "string"
      seasonPoints:
        type: "string"
      pickNumber:
        type: "integer"
        format: "int64"
      round:
        type: "integer"
        format: "int64"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"