# src/model/draftResult.ts
interface DraftResult {
  year: number,
  teamName: string,
  manager: string,
  playerName: string,
  pick: number,
  round: number,
  pointsScored: number,
  teamId: string,
  playerId: string,
  managerId: string,
  playerPosition: string
}